Start dealing with time signatures.
I don't know whether time signature should be a per-sequencer,
per-project, or per-clock concept. Per clock is probably too rigid,
as that implies a singleton for the whole song. LFOs are an example
of something that needs to know about beats but don't need to be
a child of a sequencer, so it's probably not per sequencer. That
leaves per project, with automation to change it during the song.