evolving thinking on usize
I keep going back and forth. I think that the right
way to think about usize is to use it when the value's
range should reasonably be unbounded. Could the
sample rate be a zillion? Can the number of beats in
a song be a zillion? Yes to those. But possible MIDI
notes? No, that's properly a u8.