iced experiments.
First impression: the examples each invent their own idiom for
using the library. None seems to simply use the library in a
clear, direct fashion. This is all aside from the fact that the examples
don't match the current library version that we pick up by using
crate 0.4.x.

If there aren't any examples showing how easy the library is to use,
but they all show different ways it's hard to use, then the
library might need a redesign.