Replace sequential-only sequencer.
The new one is theoretically random-access, so that we should be able to
replay songs from the GUI without setting up the Orchestrator all over again.
Skip to random points in the song, too.

This code is still scary. It's fragile and exhibits lots of f32 precision
problems.

This CL is messy. Cleanup is next.