Sources/SinksControl -> Sources/SinksUpdates
Big change. No more Controllers.
WatchedClock handles messages.

Also figured out why I had to do clone/downgrade in two lines.
The compiler can't deal with Rc::clone of one type, then upcast
to another type. Rather, it assumes the Rc::clone is for the upcast
type, which fails. So if I throw in a turbofish for the source type,
it is able to recognize the need for the cast. It's more work for
me because I have to type out an explicit type, but I have fewer
temporary variables this way, so I prefer it.