Delete WatchesClock.
A tribute to this trait: it's the one that made me realize how Rust traits are
supposed to work. A "thing" is defined by its behavior, rather than what it
actually is. If you reduce the bulk of the code's relationship with a struct to
the traits, and leave the struct-specific impl stuff to setup if possible, then
suddenly the code gets a lot cleaner because you really don't care *what* a
thing is; you only care that it honors a certain trait. WatchesClock didn't
survive, but it was the trait that allowed me to envision today's
instrument/controller/effect trait bundles.