Successfully dispatching messages.
That said, they're going to the wrong place, but the code doesn't know that
because the messages are structured wrong. I think what I'm going to do is
define messages for all the entities, but wrap them on the way up in something
like an EntityMessage(uid, message), so that when Orchestrator gets them during
an update, it can look up the right entity easily, then hand the message to it.