Replace custom test code with scripts.
Also included a script to eval all the sound scripts. If you run
`./scripts/calculate-checksums > checksums.txt` from the root of the source
directory, you'll regenerate all the scripts, as well as checksums to see
whether any of the output wav files changed. They might change for good reasons,
but this will be a tickle to have a listen and see what's going on.