Bargain-basement reverb.
The allpass delay of this Schroder reverb is broken, so I'm disabling those two
delay lines for now. They seem correct, but during setup they're calculating a
decay factor that is so close to zero that it's zero, which means the sound dies
going through them. I am undoubtedly misunderstanding the constants that go into
the calculations.

Fortunately, even bad reverb can sound good. So it's a step forward.

Thanks to BasicSynth (Daniel R. Mitchell) for providing the Schroeder
algorithm.