Sample rate shouldn't be a project attribute
Fixes #57 and maybe fixes #24

Two issues, neither of which is well-addressed with this commit.

1. Sample rate isn't specified by the project; it's what the rendering
medium wants. So an MP3 might be based on a 44.1KHz rate, but
a Windows PC might have a 48KHz audio-output stream. The
project must be rendered according to the output's needs.

2. I think Linux just happened to enumerate mono streams first,
and I got lucky that they matched what I was providing. Or
maybe Windows doesn't provide mono stream options at all.
Either way, I needed to throw in a quick hack to double-up
the mono sample if the config is stereo.