Fix more envelope/arp bugs
This code is bad and should be scrapped. Problems:

- The envelope code works, but it's more complicated than it needs to be. It
keeps around a bunch of extra steps that aren't relevant to anything. I think I
did this originally for some misguided notion of random-access accessibility,
but I've come to understand that having that ability is so rare that we can't
build anything useful on top of it. Samples, oscillators, simple effects, and
certain controllers like sequencers can give an amplitude for any given point in
time, but everything else -- envelopes, filters, arpeggiators, and anything
built with envelopes or stateful effects -- reacts to input, and it simply
doesn't make sense for them to be rewindable. I think a better approach is for
in-the-moment devices to assume forward linear movement of time, and deal
properly with inputs that could include a clock reset.

 - The sequencer thinks in terms of MIDI events, which makes it very hard to
 reconstruct what should happen when, for example, the input key is lifted and
 the arpeggiator should stop playing. Intuitively, notes that are on should be
 turned off. But if we're not tracking which note-off belongs to which note-on,
 then this is hard. For now, we're maintaining a crude table of which notes are
 on, plus which channels were sent the note-on. If we stored note events as a
 time + duration, then it would be much easier to clean up when we need to, and
 it would eliminate the possibility that the caller adds an on but forgets the
 off.