Starting to eliminate Updateable.
When I started to grok the Elm architecture, it seemed attractive to migrate
entities to messages. But there have been enough problems with Elm and Rust
fighting with each other that I don't think it's worth it in most cases. Every
entity has had to take on the responsibility of handling messages, even if they
never receive any. Most of these messages end up manipulating public
getters/setters, so they're just a more complicated way of calling those
methods.

It would have been different if I'd been able to figure out how to get the
entities themselves to handle the Iced view() method. But I couldn't; there were
just too many issues with the borrow checker (basically, either you're using
interior mutability, or you're protecting them with a mutex, and in either case
the receiver is temporary, which gives the borrow checker fits).

For now, the plan is as follows:

- Keep update() for MIDI messages.
- Continue building out the GUI app to handle EntityMessage on behalf of the
entities.
- Migrate the Tick message to a separate method/trait, maybe part of Terminates.
- Change the Updateable trait into a MIDI-only thing, as I believe that's all
that will be left.