Tricky testing
It was hard to get attack/release to respond immediately, but also
have other states end precisely on time and at their expected values.
I ended up introducing a "fast response" path that started moving
the amplitude in the current frame, and necessarily calculating the
rate of change to include that extra frame. Seems to work.