VoiceStore was handing out just-reserved voices
When we moved to enqueueing events like note on/off, we
didn't quite migrate everyone that was paying attention to
those states. So if we asked for two voices in the same tick,
the voice store would give out one, then not realize that it
was no longer available. So we had to teach it about pending
events. This entailed a lot of cleanup.