IsInstrument: HandlesMidi
Arpeggiator is turning out to be a problem. It's an instrument because it
responds to MIDI, but it's also a controller because it produces events.
It's the only instrument that doesn't produce audio.

Maybe it's a silent IsInstrument as well as an IsController. That
probably breaks a bunch of code.....