StoresVoices returns voice iters
We needed this to be able to adjust concrete voice parameters from the
#[controllable] synth params.

Goal was to start building FmSynthesizer's GUI view, which was harder
than it could be, I think. But I am starting to make parameter types
more specific, and that's working out fairly well.

We were also able to eliminate the "Uh-oh, StoresVoices is turning into a
synth" comment; `set_pan()` is gone, and anyone who needs that can
instead use voices_mut().