Progress on #54: low CPU when idle
With the benefit of experience, I was able to use mpsc channels as
designed and avoid a bunch of stupid signaling and busy-waiting.

While I was in there, I also resolved rid of various TODOs about
MIDI ports in the GUI. I still have more cleanup to do, but it is better
and apparently not broken, so I'm going to checkpoint.