Fix #54: idle CPU usage improved
Instead of the audio engine racing to keep the buffer full, we now wait
for the audio callback to determine that it needs more data. This
seems like it should be less efficient, but it actually caused us to
figure out that the callback asks for 2.6K samples, and we were
generating only 64 at a time. Now that we understand the supply/
demand situation better, we redesigned to do the right amount of
work at the right time, which seems to be better in every respect.

I can still hear some static that I believe is either under- or
over-runs of the ring buffer. I'll get to that.