bye cargo-bundle.
Didn't realize the MSI support was just broken (README says it's
experimental, but apparently it's been broken for a long time).
It seems the emphasis of that project is actually OS X, so it's
pretty much completely disjoint from what I need.

But I do now know how to upload .deb to release assets, so I'm
going to see if I can do that.