Get functional tests working again with --debug
If you pass `--debug` to the CLI, it will load assets relative to the
current working directory. Otherwise it loads from a global directory
that gets created by the installer.

This isn't a great solution, but it allows us to check in code that
works in both dev and prod cases. Eventually we should be checking
all these directories for resources, and we need a stable path system
to allow projects to refer to global, user-specific, and dev resources
simultaneously