Added EnvelopeParams as #[nano(control=false)]
As we make progress toward #122, we need the ability to specify
certain #[nano] fields that can be handled by update(), but aren't
expected to respond to F32ControlValue. An example is EnvelopeParams,
which could be flattened into ADSR, but is better as a serializable
thing that isn't a controllable thing. This meant teaching #[nano]
that some fields should populate the Nano params struct, and
should be handled in update(), but should not appear in the
automation GUI or respond to F32ControlValue (which makes no
sense <EnvelopeParams>::into()).

- Taught #[nano] to be smarter
- Moved ToySynth -> DebugSynth, and created a more standard ToySynth
- Moved all the proc-macro tests from struct-sync into the orchestration
  crate, so that they'll be tracked with this repo and tested.