Improve file loading
Remarkably, this work is to enable
https://github.com/sowbug/groove/issues/124, which just wants to display
any error encountered when a project loads. To do that, I needed to
create a bad test project, and that led to adding more flexibility in
specifying the thing browser's paths. I could have fixed 124 without
this work, but the bad behavior has been grating for a while -- it
basically broke all the official builds.

#124 is still not fixed.