Fix #107: we're responsible for note-off events
This is a kind of half-assed fix. I made MidiNoteMinder, which tracks
which notes are still active, and generates a bunch of MIDI commands to
shut them off when asked. The more industrial-strength solution is for
individual controllers not to have to worry about this, and for
Orchestrator to take care of it for everyone. I'll get to it.