Fix design error with app messages
For speed, I originally let the app define the
messages that the services sent to it. This meant
that no example code could use the services
without bringing in the app's entire Message definition.
Refactored so that each service has its own set of
messages, as well as its own channel for sending them,
and the app now checks each channel rather than a
single all-purpose one.