Got rid of weird + action in detail
Replaced with weird + action in menu
Palette is still disabled

I tried to get EntityFactory out of Orchestrator,
but I then figured out that it's better for it to be
able to instantiate by Key, because that means that
an Entity doesn't have to be Send (picture an
add_entity(entity: Entity) method on Orchestrator,
then make that a channel action, and you see
what I mean).

We can still instantiate entities elsewhere, e.g. for
testing.