Fix #127 - BPM is automatable
This was indeed a bit tricky. ControlRouter used to take a ThingStore,
but Transport (which owns Tempo) doesn't live in a ThingStore. If I gave
Orchestrator a ThingStore, then Transport's concrete type would be
erased, which is bad. I couldn't figure out how to pass a ThingStore in
the route() callback in a way that the borrow checker liked, so the
callback has evolved into a "please update (uid, index, value)" request rather than
"please lend me (uid)" or "please give me a way to get uids by myself."

Learned about some latent warts with how I'm using Serde. I keep
resisting letting it serialize everything because I like distinguishing
the real persistent state from the ephemeral housekeeping. But I
haven't developed an idiom for restoring the housekeeping
post-deserialization. So bits and pieces of things are broken. Might be
better to take a deep breath, give in to the Serde, and start without
any #[serde(skip)]s. Then carefully skip things as needed.