Improving public API
- Got rid of EntityFactory in Orchestrator. The app can manage all that.
- Introduced SelectionSet, which will help migrate away from selections in the
API.

I instantly hated the RPC-like API that selects things and then does stuff on
the selection. It's stateful and feels inflexible and clunky. It might be
right for the menu-based GUI, but I'd prefer that the GUI have the ability
to manage selection sets on its own, rather than making it an intrinsic
part of the API. We'll see how far this gets.