Starting to understand how Actions should work
lots of YAGNI cleanup

An Action should be part of the API between a UI generator (show,
ui_content, etc.) and whoever called it. It means "I can't handle this
myself, so please take care of it for me." So, for example,
OrchestratorAction contains a ClickTrack action because OrchestratorPanel
(which owns Orchestrator) handles track selection.

(The reason owners of collections manage selection rather than the
collections is that the collections have tended to be relatively simple things that
don't need to know about selection. Not sure if that's going to continue
making sense.)