Developed an opinion about UI!
Each track displays an arrangement and a device view.

The arrangement is as we've already designed: small view of audio, midi,
automation, etc.

The device view is small, medium or large.

Small is short (128x16 or so) rectangles that represent the chain of audio flow
in the track, with each providing only the most critical status/info -- maybe
its name and whether signal is flowing.

Medium is taller (128x64 or so) and contains editable knobs.

Large is windowed, so it can be as big as it wants. It's akin to a VST window.

Small is active on every track (maybe).

Medium is available only on the current track. Double-click the header to open
it, and to close anyone else's.

Big is invoked by clicking something on a single device's medium view.

For automation, sequencing, WAV editing, etc., I'm still figuring that out.