Got rid of ui_small and ui_large
I'm going to try letting the implementers figure it out from
ui.available_size(). Otherwise there needs to be shared knowledge about
constants, or an API change to provide them... and that's basically
ui.available_size().

If someone wants to be super-clever and make something that always
resizes perfectly, great.