DragDropManager loses its lock
Instead of giving out clones of Arc<Mutex<>> of DDM, we're keeping a single copy
in MiniDaw and handing out references during the egui show() step. This mattered
because of an issue trying to serialize an Arc<Mutex<>> for the upcoming
PianoRoll.

The issue is that Serde will happily serialize two of something that's supposed
to be a single thing shared between two structs. You get around this by
declaring all but one to be #[serde(skip)]. But that raises the problem how to
relink them to the original, correct one. You have to get your
post-deserialization story straight. As I was starting to do that by hooking
`Serializable::after_deser()`, but it got complicated, and I decided it was
cleaner not to give everyone a long-lived clone in the first place, but instead
to hand it out as a method parameter in `show()`. This seems to be working.

(Serde makes sure all this happens explicitly by disabling serialization of
Arc<> unless you've enabled the `rc` feature, which you hopefully won't do
without reading the warning in the
[docs](https://serde.rs/feature-flags.html#rc)).