New example Explorer
Meant to show widgets in an isolated context. I'm using it to prototype a
solid drag-and-drop system.

The rest of the code is in a bit of a mess, mostly because I've been prototyping
and duplicating a lot of existing code, so it's getting confusing which parts
are live and which aren't. To resolve this, I created the explorer as a sandbox
for this work, and then I'll have a better place to put scraps than the
production code. Stay tuned.