Performs is now part of Controls
To handle zero-length performances, we needed a simple way to catch when we're
already done before starting. I could have added something like
"update_is_finished()" into the Controls trait, but I noticed that
is_performing() and is_finished() seemed to be very similar (they might actually
be strict opposites), and that is_performing() got set by play(), but play() was
in a different trait from is_finished(). I then noticed that almost everyone who
implements Controls also implements Performs. So by combining them, I could add
a contract term that is_finished() must be correct after play().

I'd hoped this would deal with the case in MiniDAW that advanced the transport
slightly every time we pressed play. Nope; it needed a little extra handling.
But it was indeed in the same neighborhood as this fix, so we fixed it, and I
don't think it broke the design.