Demo sidechaining
As expected, this exposed a few issues.

- MIDI was over-routing. If any sequencer emitted a note, everyone would play
  it. See #145 for more.
- It's unclear how to handle hybrid controller/effect combos such as
  SignalPassthroughController. As-is, any event that is triggered by effects
  processing won't get dispatched until the next spin of the main loop. This
  means that during live play, a sidechain (for example) wouldn't trigger until
  64 samples after the source signal fires. That's a delay of 1.45 milliseconds
  @ 44.1KHz. Unacceptable. See #146 for more.

Aside from these issues, this is a really awful sidechain, because the ducking
mechanism perfectly follows the source waveform, which basically means that it's
duplicating the inverted waveform on the ducked track. We need to put an
envelope on it or something. "It" could be the passthrough controller, or the
gain, or something in between. Not sure.