Fix #139: EntityFactory uses OnceCell
This is a WIP implementing an egui component for ControlAtlas. This work caused
my acceptance (as in stage 5 of grief) that the sharing of EntityFactory was
getting out of control. Fortunately, EntityFactory seems to be a perfect
candidate for the once_cell crate, as it's initialized once at the start of the
program and is never written to or reinitialized.

The ControlAtlas GUI component is incomplete. Adds/removes ControlTrips.