Displays::ui() now returns a Response
For the most part, I was able to return a meaningful Response from every former
show() implementation. Nonsensical ones are marked with ui.label("TODO").

I'm still not sure whether a trait DisplaysInTimeline needs to exist. I'm fairly
sure it doesn't for that specific case, because we can easily pass an immutable
view_range before calling plain ui(). But mutable references like Orchestrator
or DragDropManager are harder, as they raise lifetime issues.

The answer, I suspect, is that custom versions of ui_with_arguments(...) are
fine, because those either didn't need to be traits at all (since they're
one-offs, or at least we'll always know the concrete struct type), or they are
not one-offs, but rather a repeating... trait... of a type, and we'll sometimes
have just the type-erased version of things implementing the trait, which means
we'll have a strong, specific case that justifies the Trait.

TL;DR: Keep Displays. Next, get rid of DisplaysInTimeline and replace with a
set_view_range() non-trait method or some such. Finally, be on the lookout for
future Displays-like traits.