#145 test that ext MIDI reaches instruments
"By default, MIDI messages from external devices are sent to all tracks.
Eventually, should be opt-in per channel per track." This tests the first part
of that. The second part is unimplemented.

While in here, I reworked ToyInstrument to have a shareable counter. This means
that we can still get information from it even after the type erasure that
happens when it's added to EntityStore.