Fix #156: Route MIDI messages to external interface
I had expected to add an OrchestratorEvent::Midi so that the app could handle it
by forwarding to the MIDI interface. But I forgot that we call
Orchestrator::render() in the NeedsAudioFn, so that's where we receive our
messages. Fortunately, it wasn't hard to get a MidiInterfaceInput channel sender
in there, so (a) I can do it at all, and (b) since it's a channel rather than a
method call, I don't have to worry about NeedsAudioFn becoming janky if
MidiInterface gets slow.